requirements.txt
cryptography==41.0.2
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.31.0